Extract LeaderboardBase component from duplicate leaderboards
Create a reusable LeaderboardBase component that handles all shared
leaderboard UI (medal colors, ranking, metric selector, list rendering).
UserLeaderboard and ForecastLeaderboard are now thin wrappers that handle
data fetching and transformation.

Also updates ForecastLeaderboard to use the useScores hook instead of
direct API calls.
